feat(credentials): add external credential resolution for TACACS+ runtime server configs
- Introduced a new library `tacacsrs-credentials` to handle external credential resolution.
- Implemented functions to resolve and validate various credential types including EPSK, raw private keys, and certificates.
- Added a `CredentialResolver` trait to define methods for resolving and validating credentials.
- Integrated the new credential resolution logic into the existing TACACS+ server configuration handling.
- Updated dependencies in `Cargo.toml` for the new library.
- Refactored existing code to utilize the new credential resolution methods, improving modularity and maintainability.
