+++ executables/tacon/Cargo.toml
@@ -19,6 +19,7 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tacacsrs-agent-client = { path = "../../libraries/tacacsrs_agent_client" }
 tacacsrs-config = { path = "../../libraries/tacacsrs_config" }
+tacacsrs-credentials = { path = "../../libraries/tacacsrs_credentials" }
 tacacsrs-messages = { path = "../../libraries/tacacsrs_messages" }
 tacacsrs-networking = { path = "../../libraries/tacacsrs_networking" }
 tokio = { version = "1.49.0", features = ["full"] }
