+++ libraries/tacacsrs_agent/Cargo.toml
@@ -13,6 +13,7 @@ async-trait = "0.1.89"
 log = "0.4.29"
 tacacsrs-agent-client = { path = "../tacacsrs_agent_client" }
 tacacsrs-config = { path = "../tacacsrs_config" }
+tacacsrs-credentials = { path = "../tacacsrs_credentials" }
 tacacsrs-messages = { path = "../tacacsrs_messages" }
 tacacsrs-networking = { path = "../tacacsrs_networking" }
 tokio = { version = "1.49.0", features = ["full"] }
