+++ libraries/tacacsrs_config/Cargo.toml
@@ -13,5 +13,8 @@ bitflags = { version = "2", features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 
+[dev-dependencies]
+tacacsrs-credentials = { path = "../tacacsrs_credentials" }
+
 [lints]
 workspace = true
