+++ libraries/tacacsrs_config/examples/quick_start.rs
@@ -1,4 +1,4 @@
-use tacacsrs_config::{parse_yang_json, resolve_servers};
+use tacacsrs_config::{enumerate_servers, parse_yang_json};
 
 fn main() -> anyhow::Result<()> {
     let json = r#"{
@@ -17,21 +17,24 @@ fn main() -> anyhow::Result<()> {
     }"#;
 
     // Parse YANG config
-    let config = parse_yang_json(json, None)?;
+    let config = parse_yang_json(json)?;
     println!("📄 Parsed one TACACS+ server from YANG JSON");
 
-    // Resolve servers (None = no external keystore/truststore resolver needed)
-    let servers = resolve_servers(&config, None)?;
-    println!("✅ Resolved runtime-ready server entries\n");
+    // Enumerate servers (bundle references are materialized inline when present)
+    let servers = enumerate_servers(&config)?;
+    println!("✅ Enumerated per-server config entries\n");
 
     for server in &servers {
         println!("  ┌─ server: '{}'", server.name);
-        println!("  │  endpoint: {}", server.socket_address());
+        println!("  │  endpoint: {}:{}", server.address, server.port);
         println!("  │  single_connection: {}", server.single_connection);
-        println!("  │  timeout: {:?}", server.timeout_duration());
+        println!("  │  timeout: {}s", server.timeout);
         println!(
             "  │  transport: {}",
-            if server.is_tls() {
+            if server.client_identity.is_some()
+                || server.server_authentication.is_some()
+                || server.hello_params.is_some()
+            {
                 "TLS"
             } else {
                 "obfuscation"
