+++ libraries/tacacsrs_config/src/mapping.rs
@@ -1,3 +1,4 @@
 // This module previously contained ServerConnectionConfig, ResolvedSecurity,
 // to_connection_configs(), and resolve_security(). These have been replaced by
-// the resolver-based API in resolvers.rs (ResolvedServer, resolve_servers, etc.).
+// the bundle-enumeration API in enumeration.rs and external credential
+// resolution in tacacsrs-credentials.
