+++ libraries/tacacsrs_config/tests/parse_validation.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 
 use tacacsrs_config::{
-    resolve_server, parse_yang_json, parse_yang_json_file, pipeline,
+    enumerate_server as resolve_server, parse_yang_json, parse_yang_json_file, pipeline,
     validate_credential_references, YangConfigRoot, TacacsPlusServerType,
 };
 
@@ -39,7 +39,7 @@ fn parse_minimal_obfuscation_config() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).expect("should parse minimal config");
+    let config = parse_yang_json(json).expect("should parse minimal config");
     assert_eq!(config.server.len(), 1);
 
     let s = &config.server[0];
@@ -70,7 +70,7 @@ fn parse_multi_type_server() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).unwrap();
+    let config = parse_yang_json(json).unwrap();
     let st = config.server[0].server_type;
     assert!(st.contains(TacacsPlusServerType::AUTHENTICATION));
     assert!(st.contains(TacacsPlusServerType::AUTHORIZATION));
@@ -115,7 +115,7 @@ fn parse_tls_config() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).unwrap();
+    let config = parse_yang_json(json).unwrap();
     let s = &config.server[0];
     assert_eq!(s.server_type, TacacsPlusServerType::ACCOUNTING);
     assert_eq!(s.domain_name.as_deref(), Some("tacacs.example.com"));
@@ -167,7 +167,7 @@ fn parse_multiple_servers_with_failover_order() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).unwrap();
+    let config = parse_yang_json(json).unwrap();
     assert_eq!(config.server.len(), 2);
     assert_eq!(config.server[0].name, "primary");
     assert_eq!(config.server[1].name, "secondary");
@@ -196,7 +196,7 @@ fn reject_duplicate_address_port() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("duplicate server address+port"), "unexpected error: {err}");
 }
 
@@ -217,7 +217,7 @@ fn reject_sni_without_domain_name() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string().contains("sni-enabled requires domain-name"),
         "unexpected error: {err}",
@@ -232,7 +232,7 @@ fn reject_empty_server_list() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("at least one entry"), "unexpected error: {err}");
 }
 
@@ -252,7 +252,7 @@ fn reject_invalid_server_type() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("unknown variant"), "unexpected error: {err}");
 }
 
@@ -272,7 +272,7 @@ fn default_timeout_applied() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).unwrap();
+    let config = parse_yang_json(json).unwrap();
     assert_eq!(config.server[0].timeout, 5);
 }
 
@@ -320,7 +320,7 @@ fn credential_references_preserved_for_roundtrip() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).unwrap();
+    let config = parse_yang_json(json).unwrap();
     let s = &config.server[0];
 
     let ci = s.client_identity.as_ref().unwrap();
@@ -352,7 +352,7 @@ fn reject_missing_credential_reference() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("nonexistent"), "unexpected error: {err}");
 }
 
@@ -371,7 +371,7 @@ fn reject_no_security() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("security requires one of [tls, obfuscation]"),
@@ -402,7 +402,7 @@ fn reject_tls_and_obfuscation() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("security allows only one of [tls, obfuscation]"),
@@ -428,7 +428,7 @@ fn reject_missing_inline_or_keystore_choice() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("client-identity/certificate requires one of [inline, central-keystore]"),
@@ -454,7 +454,7 @@ fn reject_missing_inline_or_truststore_choice() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string().contains(
             "server-authentication/ca-certs requires one of [inline, central-truststore]"
@@ -486,8 +486,8 @@ fn accept_tls13_epsk_config_without_runtime_psk_support() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None)
-        .expect("tls13-epsk config should parse without runtime PSK support");
+    let config =
+        parse_yang_json(json).expect("tls13-epsk config should parse without runtime PSK support");
     assert_eq!(config.server.len(), 1);
     assert_eq!(config.server[0].name, "epsk-server");
 }
@@ -532,7 +532,7 @@ fn parse_yang_json_file_parses_and_validates_config() {
     }"#;
 
     let path = write_temp_json_file(json);
-    let result = parse_yang_json_file(&path, None);
+    let result = parse_yang_json_file(&path);
     let _ = std::fs::remove_file(&path);
 
     let config = result.expect("file-based parse should succeed");
@@ -590,11 +590,10 @@ fn resolve_server_finds_server_by_name() {
                 ]
             }
         }"#,
-        None,
     )
     .expect("config should parse");
 
-    let result = resolve_server(&config, "exists", None);
+    let result = resolve_server(&config, "exists");
     assert!(result.is_ok(), "expected successful lookup");
 }
 
@@ -614,11 +613,10 @@ fn resolve_server_rejects_unknown_name() {
                 ]
             }
         }"#,
-        None,
     )
     .expect("config should parse");
 
-    let err = resolve_server(&config, "missing", None).unwrap_err();
+    let err = resolve_server(&config, "missing").unwrap_err();
     assert!(err.to_string().contains("server 'missing' not found"), "unexpected error: {err}",);
 }
 
@@ -638,11 +636,10 @@ fn validate_credential_references_returns_ok_for_simple_config() {
                 ]
             }
         }"#,
-        None,
     )
     .expect("config should parse");
 
-    validate_credential_references(&config, None)
+    validate_credential_references(&config)
         .expect("validate_credential_references should return Ok for simple config");
 }
 
@@ -662,7 +659,7 @@ fn reject_empty_server_type_bitflags() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("at least one bit must be set"), "unexpected error: {err}",);
 }
 
@@ -686,7 +683,7 @@ fn reject_duplicate_client_credentials_ids() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("duplicate client-credentials id"));
 }
 
@@ -710,7 +707,7 @@ fn reject_duplicate_server_credentials_ids() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("duplicate server-credentials id"));
 }
 
@@ -732,7 +729,7 @@ fn reject_missing_client_credential_reference() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("credentials-reference 'missing-client' not found"),
@@ -758,7 +755,7 @@ fn reject_raw_private_key_without_choice() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("client-identity/raw-private-key requires one of [inline, central-keystore]"),
@@ -789,7 +786,7 @@ fn reject_raw_private_key_with_multiple_choices() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string().contains(
             "client-identity/raw-private-key allows only one of [inline, central-keystore]"
@@ -816,7 +813,7 @@ fn reject_raw_public_keys_without_choice() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string().contains(
             "server-authentication/raw-public-keys requires one of [inline, central-truststore]"
@@ -848,7 +845,7 @@ fn reject_raw_public_keys_with_multiple_choices() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string().contains(
             "server-authentication/raw-public-keys allows only one of [inline, central-truststore]"
@@ -875,7 +872,7 @@ fn reject_source_ip_and_source_interface_together() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("source-type allows only one of [source-ip, source-interface]"),
@@ -918,7 +915,7 @@ fn reject_client_identity_reference_and_explicit_together() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("client-identity allows only one of [ref, explicit/auth-type]"),
@@ -970,7 +967,7 @@ fn reject_server_auth_reference_and_explicit_together() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("server-authentication allows only one of [ref, explicit]"),
@@ -1002,7 +999,7 @@ fn accept_server_authentication_with_ee_certs_only() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).expect("ee-certs explicit mode should be accepted");
+    let config = parse_yang_json(json).expect("ee-certs explicit mode should be accepted");
     assert_eq!(config.server.len(), 1);
     assert_eq!(config.server[0].name, "ee-only");
 }
@@ -1039,7 +1036,7 @@ fn reject_client_credentials_with_multiple_auth_types() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("client-credentials/auth-type allows only one of [certificate, raw-public-key, tls13-epsk]"),
@@ -1074,7 +1071,7 @@ fn accept_client_credentials_with_certificate_auth_type() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).expect("certificate auth-type should be accepted");
+    let config = parse_yang_json(json).expect("certificate auth-type should be accepted");
     assert_eq!(config.client_credentials.len(), 1);
     assert_eq!(config.client_credentials[0].id, "cert-only");
 }
@@ -1105,7 +1102,7 @@ fn accept_client_credentials_with_raw_private_key_auth_type() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).expect("raw-private-key auth-type should be accepted");
+    let config = parse_yang_json(json).expect("raw-private-key auth-type should be accepted");
     assert_eq!(config.client_credentials.len(), 1);
     assert_eq!(config.client_credentials[0].id, "rpk-only");
 }
@@ -1137,8 +1134,7 @@ fn accept_client_credentials_with_tls13_epsk_auth_type() {
         }
     }"#;
 
-    let config =
-        parse_yang_json(json, None).expect("tls13-epsk auth-type should be accepted with psk");
+    let config = parse_yang_json(json).expect("tls13-epsk auth-type should be accepted with psk");
     assert_eq!(config.client_credentials.len(), 1);
     assert_eq!(config.client_credentials[0].id, "epsk-only");
 }
@@ -1402,7 +1398,7 @@ fn reject_tls_version_min_below_13() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err
         .to_string()
         .contains("minimum TLS version must be >= 1.3"));
@@ -1435,7 +1431,7 @@ fn reject_tls_version_max_below_13() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err
         .to_string()
         .contains("maximum TLS version must be >= 1.3"));
@@ -1469,7 +1465,7 @@ fn accept_tls_versions_at_or_above_13() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).expect("tls version bounds should be accepted");
+    let config = parse_yang_json(json).expect("tls version bounds should be accepted");
     assert_eq!(config.server.len(), 1);
     assert_eq!(config.server[0].name, "tls-bounds-ok");
 }
@@ -1490,7 +1486,7 @@ fn accept_tls_hello_params_without_tls_versions() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).expect("empty hello-params should be accepted");
+    let config = parse_yang_json(json).expect("empty hello-params should be accepted");
     assert_eq!(config.server.len(), 1);
     assert_eq!(config.server[0].name, "tls-empty-hello");
     assert!(config.server[0].hello_params.is_some());
@@ -1519,7 +1515,7 @@ fn accept_tls13_epsk_when_present() {
         }
     }"#;
 
-    let config = parse_yang_json(json, None).expect("tls13-epsk should parse");
+    let config = parse_yang_json(json).expect("tls13-epsk should parse");
     assert_eq!(config.server.len(), 1);
     assert_eq!(config.server[0].name, "epsk-ok");
 }
@@ -1548,7 +1544,7 @@ fn reject_tls13_epsk_with_multiple_choice_sources() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).expect_err("tls13-epsk with multiple choices must fail");
+    let err = parse_yang_json(json).expect_err("tls13-epsk with multiple choices must fail");
     assert!(
         err.to_string()
             .contains("client-identity/tls13-epsk allows only one of [inline, central-keystore]"),
@@ -1597,7 +1593,7 @@ fn accept_valid_private_key_format() {
         }
     }"#;
 
-    parse_yang_json(json, None).expect("valid key formats should be accepted");
+    parse_yang_json(json).expect("valid key formats should be accepted");
 }
 
 #[test]
@@ -1635,7 +1631,7 @@ fn reject_invalid_private_key_format() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("invalid private-key-format 'bogus-format'"),
@@ -1680,7 +1676,7 @@ fn reject_invalid_public_key_format() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("invalid public-key-format"), "unexpected error: {err}",);
 }
 
@@ -1724,7 +1720,7 @@ fn reject_invalid_symmetric_key_format() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("invalid key-format 'bad-format'"), "unexpected error: {err}",);
 }
 
@@ -1768,7 +1764,7 @@ fn accept_valid_symmetric_key_format() {
         }
     }"#;
 
-    parse_yang_json(json, None).expect("valid symmetric key format should be accepted");
+    parse_yang_json(json).expect("valid symmetric key format should be accepted");
 }
 
 // ---------------------------------------------------------------------------
@@ -1807,7 +1803,7 @@ fn reject_invalid_base64_in_private_key() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("cleartext-private-key contains invalid base64"),
@@ -1839,7 +1835,7 @@ fn reject_invalid_base64_in_cert_data() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("certificate 'ca1' contains invalid base64"),
@@ -1879,7 +1875,7 @@ fn reject_empty_private_key() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("cleartext-private-key must not be empty"),
@@ -1911,7 +1907,7 @@ fn accept_pem_certificate_data() {
         }
     }"#;
 
-    parse_yang_json(json, None).expect("PEM certificate data should be accepted");
+    parse_yang_json(json).expect("PEM certificate data should be accepted");
 }
 
 #[test]
@@ -1938,7 +1934,7 @@ fn reject_pem_without_end_marker() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("PEM BEGIN marker but no END marker"),
@@ -1985,7 +1981,7 @@ fn reject_invalid_base64_in_symmetric_key() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("cleartext-symmetric-key contains invalid base64"),
@@ -2020,7 +2016,7 @@ fn reject_hidden_private_key_in_server_certificate() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("hidden-private-key"), "unexpected error: {err}",);
 }
 
@@ -2050,7 +2046,7 @@ fn reject_encrypted_private_key_in_server_certificate() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("encrypted-private-key"), "unexpected error: {err}",);
 }
 
@@ -2076,7 +2072,7 @@ fn reject_hidden_private_key_in_raw_private_key() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("hidden-private-key"), "unexpected error: {err}",);
 }
 
@@ -2103,7 +2099,7 @@ fn reject_hidden_symmetric_key_in_epsk() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("hidden-symmetric-key"), "unexpected error: {err}",);
 }
 
@@ -2133,7 +2129,7 @@ fn reject_encrypted_symmetric_key_in_epsk() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("encrypted-symmetric-key"), "unexpected error: {err}",);
 }
 
@@ -2161,7 +2157,7 @@ fn reject_epsk_context_derivation() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("context"), "unexpected error: {err}",);
 }
 
@@ -2189,7 +2185,7 @@ fn reject_epsk_target_protocol() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("target-protocol"), "unexpected error: {err}",);
 }
 
@@ -2217,7 +2213,7 @@ fn reject_epsk_target_kdf() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("target-kdf"), "unexpected error: {err}",);
 }
 
@@ -2248,7 +2244,7 @@ fn reject_hidden_private_key_in_client_credentials() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("hidden-private-key"), "unexpected error: {err}",);
 }
 
@@ -2281,7 +2277,7 @@ fn reject_encrypted_private_key_in_client_credentials_rpk() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("encrypted-private-key"), "unexpected error: {err}",);
 }
 
@@ -2313,7 +2309,7 @@ fn reject_epsk_context_in_client_credentials() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("context"), "unexpected error: {err}",);
 }
 
@@ -2348,7 +2344,7 @@ fn reject_invalid_key_format_in_client_credentials_rpk() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("invalid private-key-format 'bogus-format'"),
@@ -2384,7 +2380,7 @@ fn reject_invalid_rpk_public_key_format_in_server_auth() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("invalid public-key-format"), "unexpected error: {err}",);
 }
 
@@ -2412,7 +2408,7 @@ fn reject_invalid_ee_cert_data() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("invalid base64"), "unexpected error: {err}",);
 }
 
@@ -2439,7 +2435,7 @@ fn reject_invalid_key_format_in_server_rpk() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(
         err.to_string()
             .contains("invalid private-key-format 'invalid'"),
@@ -2471,6 +2467,6 @@ fn reject_invalid_symmetric_key_format_in_server_epsk() {
         }
     }"#;
 
-    let err = parse_yang_json(json, None).unwrap_err();
+    let err = parse_yang_json(json).unwrap_err();
     assert!(err.to_string().contains("invalid key-format 'bogus'"), "unexpected error: {err}",);
 }
