+++ libraries/tacacsrs_credentials/src/epsk.rs
@@ -1,17 +1,8 @@
-//! TLS 1.3 External Pre-Shared Key (EPSK) resolution and validation.
-//!
-//! Handles `central-keystore-reference` for `tls13-epsk` client identity.
-
 use anyhow::{Context, Result};
+use tacacsrs_config::Tls13Epsk;
 
-use crate::generated::tacacs_plus::Tls13Epsk;
+use crate::CredentialResolver;
 
-use super::CredentialResolver;
-
-/// Resolves an EPSK keystore reference to inline symmetric key material.
-///
-/// Calls [`CredentialResolver::resolve_symmetric_key`] to fetch the symmetric
-/// key material and format identity.
 pub(crate) fn resolve_epsk_keystore_ref(
     epsk: &mut Tls13Epsk,
     resolver: &dyn CredentialResolver,
@@ -26,7 +17,7 @@ pub(crate) fn resolve_epsk_keystore_ref(
                 )
             })?;
 
-        epsk.inline_definition = Some(crate::generated::keystore::SymmetricKeyInlineDefinition {
+        epsk.inline_definition = Some(tacacsrs_config::keystore::SymmetricKeyInlineDefinition {
             key_format: material.key_format.map(|f| f.as_rfc7951_str().to_owned()),
             cleartext_symmetric_key: Some(material.cleartext_symmetric_key),
             hidden_symmetric_key: None,
@@ -37,18 +28,17 @@ pub(crate) fn resolve_epsk_keystore_ref(
     Ok(())
 }
 
-/// Validates EPSK keystore references in the client-identity subtree.
 pub(crate) fn validate_epsk_keystore_refs(
-    ci: &crate::generated::tacacs_plus::TlsClientClientIdentity,
+    ci: &tacacsrs_config::TlsClientClientIdentity,
     server_name: &str,
     resolver: &dyn CredentialResolver,
     errors: &mut Vec<String>,
 ) {
     if let Some(ref epsk) = ci.tls13_epsk {
         if let Some(ref ks_ref) = epsk.central_keystore_reference {
-            if let Err(e) = resolver.validate_symmetric_key(ks_ref) {
+            if let Err(error) = resolver.validate_symmetric_key(ks_ref) {
                 errors.push(format!(
-                    "server '{server_name}': tls13-epsk central-keystore-reference: {e}",
+                    "server '{server_name}': tls13-epsk central-keystore-reference: {error}",
                 ));
             }
         }
